directory rename, and additional audio file
